pyproject.toml
[tool.poetry]
name = "investr"
version = "0.1.0"
description = ""
authors = ["niyyou <niyyou@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.11"
streamlit = "^0.86.0"
numpy = "^1.19.2"
pandas = "^1.1.2"
pyngrok = "^4.1.13"
black = "^20.8b1"
python-box = "^5.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = 'investr.run:cli'
